factor more data into zingo-testdata
