+++ zingo-testdata/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+zingoconfig = { path = "../zingoconfig" }
+
+zcash_primitives = { workspace = true }
